fix: рефакторит сборку объекта избранных предложений
